+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to Calendar Versioning (CalVer).
 ### Removed
 
 ### Fixed
+- Fix deleting all app data reusing a closed Whitenoise database instance.
 
 ### Security